fix(review): address review findings from #226
- Update R39 spec to match actual GC report shape ({ship,execute:{deleted[],kept[]}} not flat per-file)
- Strengthen --steps cleanup rejection test with javascript assertion checking error message content
- Lift listBranches() and readTtlDaysFromConfig() to lib/state.js; remove verbatim duplicates from state/ship.js and state/execute.js
- Remove redundant fs2/path2 aliases in state/ship.js::cmdGc dry-run path
