+++ docs/specs/ship-sdlc.md
@@ -70,8 +70,8 @@
   - Acceptance: starting `/ship-sdlc` on `main`, then having execute-plan-sdlc create branch `fix/foo`, results in exactly one state file named `ship-fix-foo-*.json` after the execute step (no orphan `ship-main-*.json`); the file's `data.branch` field equals `"fix/foo"`.
 - R38: Terminal `cleanup` pipeline step (issue #223). Every ship pipeline run MUST include a deterministic terminal `cleanup` step appended after every other configured step (including `pr`, `archive-openspec`, `learnings-commit`). The step is added to the resolved step list by `skill/ship.js::computeSteps`, NOT by user configuration — it is rejected with a validation error if it appears in CLI `--steps` or `ship.steps[]`. The step has `status: "will_run"` on success paths and `skill: null` (dispatched as a direct `Bash` call to `state/ship.js cleanup-pipeline`, not as an Agent). Behavior on success paths: (a) validate pipeline contract (no `pending`/`in_progress` steps among the configured set); (b) on contract success, delete the current run's state file; (c) GC sweep stale `ship-*` AND `execute-*` state files older than `state.gc.ttlDays` (default 7) whose branch no longer exists in `git branch --list`; (d) emit one JSON report `{currentRun: {valid, cleaned}, gc: {ship: {...}, execute: {...}}}`. On failure paths (any earlier step ended in `status: "failed"`), the skill MUST invoke the same script with `--force`, which skips the contract check and preserves the current run's state file (so `--resume` works) but still runs the GC sweep against unrelated stale orphans.
   - Acceptance: a clean `/ship-sdlc` end-to-end leaves `.sdlc/execution/` empty afterwards; the dry-run output table includes `cleanup` as the last step; `--steps cleanup` (or `ship.steps: ["cleanup"]`) is rejected with a validation error from `skill/ship.js`; on failure paths the current run's state file is preserved and unrelated stale orphans are still pruned.
-- R39: `/ship-sdlc --gc` on-demand pruning (issue #223). When `--gc` is passed, `skill/ship.js` MUST short-circuit pipeline composition and emit `{action: "gc", report: {...}, errors: [], warnings: []}` to the prepare-output file. The report MUST list each candidate state file with `{prefix, file, branch, reason, deleted}` where `prefix ∈ {"ship", "execute"}`. Pruning rule: a file is deleted iff (mtime is older than the TTL) AND (parsed branch slug does not match any current branch from `git branch --list`); files newer than the TTL are kept regardless (in-flight pipelines must not be wiped). The TTL defaults to `state.gc.ttlDays` from `.claude/sdlc.json` (fallback 7); CLI `--ttl-days <N>` overrides config. The same `gcStateFiles` helper from `lib/state.js` powers both `--gc` and the terminal cleanup step (R38). The skill MUST print the report and stop — no pipeline steps run.
-  - Acceptance: `/ship-sdlc --gc` produces a prepare output with top-level `action: "gc"` and a `report` array; `/ship-sdlc --gc --ttl-days 0` removes every state file whose branch is absent from `git branch --list` (regardless of age) but never removes files for currently-existing branches; no pipeline step is dispatched.
+- R39: `/ship-sdlc --gc` on-demand pruning (issue #223). When `--gc` is passed, `skill/ship.js` MUST short-circuit pipeline composition and emit `{action: "gc", report: {...}, errors: [], warnings: []}` to the prepare-output file. The report shape is `{ttlDays: N, ship: {deleted: [...], kept: [...]}, execute: {deleted: [...], kept: [...]}}` where each entry in `deleted`/`kept` is `{file, prefix, branch, reason}` (prefix ∈ `{"ship", "execute"}`). Pruning rule: a file is deleted iff (mtime is older than the TTL) AND (parsed branch slug does not match any current branch from `git branch --list`); files newer than the TTL are kept regardless (in-flight pipelines must not be wiped). The TTL defaults to `state.gc.ttlDays` from `.claude/sdlc.json` (fallback 7); CLI `--ttl-days <N>` overrides config. The same `gcStateFiles` helper from `lib/state.js` powers both `--gc` and the terminal cleanup step (R38). The skill MUST print the report and stop — no pipeline steps run.
+  - Acceptance: `/ship-sdlc --gc` produces a prepare output with top-level `action: "gc"` and a `report` object containing `ship` and `execute` sub-reports each with `deleted[]` and `kept[]` arrays; `/ship-sdlc --gc --ttl-days 0` removes every state file whose branch is absent from `git branch --list` (regardless of age) but never removes files for currently-existing branches; no pipeline step is dispatched.
 
 > **Implementation surface for R37–R39:** `state/ship.js` exposes `gc`, `migrate`, and `cleanup-pipeline` subcommands; `state/execute.js` exposes `gc`. The shared logic lives in `lib/state.js::gcStateFiles` and `lib/state.js::migrateBranchSlug`.
